backend/requirement.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy
redis
aioredis
python-dotenv
authlib
httpx
itsdangerous
alembic
python-jose[cryptography]
passlib[bcrypt]
crud
requests
pybase64
apscheduler
pytest
python-multipart
